refactor(console): cleanup `console` module, and move node-like util functions to the new `util` builtin module
